finish section

+++ components/data-table.tsx
@@ -26,6 +26,7 @@ import {
 
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
+import { useConfirm } from "@/hooks/use-confirm"
 
 interface DataTableProps<TData, TValue> {
     columns: ColumnDef<TData, TValue>[]
@@ -42,6 +43,10 @@ export function DataTable<TData, TValue>({
     onDelete,
     disabled,
 }: DataTableProps<TData, TValue>) {
+    const [ConfirmDialog, confirm] = useConfirm(
+        "Are you sure?",
+        "You are about to delete this transaction."
+    );
     const [sorting, setSorting] = React.useState<SortingState>([])
     const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
         []
@@ -67,6 +72,7 @@ export function DataTable<TData, TValue>({
 
     return (
         <div>
+            <ConfirmDialog />
             <div className="flex items-center py-4">
                 <Input
                     placeholder={`Filter ${filterKey}...`}
@@ -82,7 +88,9 @@ export function DataTable<TData, TValue>({
                         size="sm"
                         variant='outline'
                         className="ml-auto font-normal text-xs"
-                        onClick={() => {
+                        onClick={async () => {
+                            const ok = await confirm();
+                            if (!ok) return;
                             onDelete(table.getFilteredSelectedRowModel().rows)
                             table.resetRowSelection()
                         }}
